Make all error reporting in __init__ follow the same form
It would be nice to move as much of this as possible to Source.__init__
but I don't see how right now. Perhaps after more code is moved up to
the Source class it can happen.

For now, everything follows the "What: error." format. "What" is the
class, the error does not start with a capital letter, and it ends with
a period.